Use Unordered comparison for NaN canonicalization
The current implementation of NaN canonicalization determines if a float
is NaN by checking if it is not equal to itself. But on x86 this results
in two jumps: one to check for orderedness and a second to check for
equality. But if a value is unordered w.r.t itself then it must be a
NaN, so it's enough to just check that which only results in one jump.